Clear C++20 enum-enum conversion warnings

+++ components/victron_ble/sensor/victron_sensor.h
@@ -222,7 +222,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0x3FFFFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(0.001f * val);
+      this->publish_state(0.001f * static_cast<int32_t>(val));
     }
   };
 
@@ -230,7 +230,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0xFFFFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(-0.1f * val);
+      this->publish_state(-0.1f * static_cast<u_int32_t>(val));
     }
   };
 
@@ -238,7 +238,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0x3FFFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(val);
+      this->publish_state(static_cast<int32_t>(val));
     }
   };
 
@@ -246,7 +246,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0x7FFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(0.01f * val);
+      this->publish_state(0.01f * static_cast<int16_t>(val));
     }
   };
 
@@ -254,7 +254,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0x7FFF) {
       this->publish_state(0.0f);
     } else {
-      this->publish_state(0.01f * val);
+      this->publish_state(0.01f * static_cast<int16_t>(val));
     }
   };
 
@@ -262,7 +262,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0xFFFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(0.01f * val);
+      this->publish_state(0.01f * static_cast<u_int16_t>(val));
     }
   };
 
@@ -270,7 +270,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0x7FFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(0.1f * val);
+      this->publish_state(0.1f * static_cast<int16_t>(val));
     }
   };
 
@@ -278,7 +278,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0xFFFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(0.1f * val);
+      this->publish_state(0.1f * static_cast<u_int16_t>(val));
     }
   };
 
@@ -286,7 +286,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0x7FFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(val);
+      this->publish_state(static_cast<int16_t>(val));
     }
   };
 
@@ -294,7 +294,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0xFFFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(val);
+      this->publish_state(static_cast<u_int16_t>(val));
     }
   };
 
@@ -302,7 +302,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0xFFFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(0.01f * val - 273.15f);
+      this->publish_state(0.01f * static_cast<u_int16_t>(val) - 273.15f);
     }
   };
 
@@ -310,7 +310,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0x7FFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(0.01f * val);
+      this->publish_state(0.01f * static_cast<u_int16_t>(val));
     }
   };
 
@@ -318,7 +318,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0x3FFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(0.01f * val);
+      this->publish_state(0.01f * static_cast<u_int16_t>(val));
     }
   };
 
@@ -326,7 +326,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0x1FFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(0.01f * val);
+      this->publish_state(0.01f * static_cast<u_int16_t>(val));
     }
   };
 
@@ -334,7 +334,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0xFFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(0.01f * val);
+      this->publish_state(0.01f * static_cast<u_int16_t>(val));
     }
   };
 
@@ -342,7 +342,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0x7FF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(0.1f * val);
+      this->publish_state(0.1f * static_cast<u_int16_t>(val));
     }
   };
 
@@ -350,7 +350,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0x3FF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(0.1f * val);
+      this->publish_state(0.1f * static_cast<u_int16_t>(val));
     }
   };
 
@@ -358,7 +358,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0x1FF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(-0.1f * val);
+      this->publish_state(-0.1f * static_cast<u_int16_t>(val));
     }
   };
 
@@ -366,7 +366,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0x1FF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(0.1f * val);
+      this->publish_state(0.1f * static_cast<u_int16_t>(val));
     }
   };
 
@@ -374,7 +374,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0x7F) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(0.01f * val + 2.60f);
+      this->publish_state(0.01f * static_cast<u_int16_t>(val) + 2.60f);
     }
   };
 
@@ -382,7 +382,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0x7F) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(val);
+      this->publish_state(static_cast<u_int16_t>(val));
     }
   };
 
@@ -390,7 +390,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val == 0x7F) {
       this->publish_state(NAN);
     } else {
-      this->publish_state(-40.0f + val);
+      this->publish_state(-40.0f + static_cast<u_int16_t>(val));
     }
   };
 
@@ -398,7 +398,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val_C == 0x3FFFFF || val_V == 0x1FFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state((0.001f * val_C) * (0.01f * val_V));
+      this->publish_state((0.001f * static_cast<int32_t>(val_C)) * (0.01f * static_cast<int16_t>(val_V)));
     }
   };
 
@@ -406,7 +406,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val_C == 0x7FFF || val_V == 0x7FFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state((0.1f * val_C) * (0.01f * val_V));
+      this->publish_state((0.1f * static_cast<int16_t>(val_C)) * (0.01f * static_cast<int16_t>(val_V)));
     }
   };
 
@@ -414,7 +414,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val_C == 0xFFFF || val_V == 0xFFFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state((0.1f * val_C) * (0.01f * val_V));
+      this->publish_state((0.1f * static_cast<u_int16_t>(val_C)) * (0.01f * static_cast<u_int16_t>(val_V)));
     }
   };
 
@@ -422,7 +422,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val_C == 0x7FFF || val_V == 0x3FFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state((0.1f * val_C) * (0.01f * val_V));
+      this->publish_state((0.1f * static_cast<int16_t>(val_C)) * (0.01f * static_cast<u_int16_t>(val_V)));
     }
   };
 
@@ -430,7 +430,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val_C == 0x7FF || val_V == 0x1FFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state((0.1f * val_C) * (0.01f * val_V));
+      this->publish_state((0.1f * static_cast<u_int16_t>(val_C)) * (0.01f * static_cast<u_int16_t>(val_V)));
     }
   };
 
@@ -438,7 +438,7 @@ class VictronSensor : public Component, public sensor::Sensor, public Parented<V
     if (val_C == 0x1FF || val_V == 0x7FFF) {
       this->publish_state(NAN);
     } else {
-      this->publish_state((-0.1f * val_C) * (0.01f * val_V));
+      this->publish_state((-0.1f * static_cast<u_int16_t>(val_C)) * (0.01f * static_cast<u_int16_t>(val_V)));
     }
   };
 };
